+++ include/navkit/app_support/emulation/EmulatorBindingPolicy.hpp
@@ -6,21 +6,20 @@
 #include "navkit/app_support/emulation/EmulatorBinding.hpp"
 #include "navkit/app_support/emulation/EmulatorPolicy.hpp"
 #include "navkit/core/estimation/sensor/SensorPolicy.hpp"
-#include "navkit/io/RunLogger.hpp"
 
 #include <concepts>
 
 namespace navkit::app_support
 {
 
-template<typename Candidate>
+template<typename Candidate, typename Logger>
 concept EmulatorBindingPolicy =
     requires {
         { Candidate::Id } -> std::convertible_to<SensorId>;
         typename Candidate::Emulator_t;
         typename Candidate::Sensor_t;
     } && navkit::core::estimation::SensorPolicy<typename Candidate::Sensor_t> &&
-    EmulatorPolicy<typename Candidate::Emulator_t, typename Candidate::Sensor_t, io::RunLogger> &&
+    EmulatorPolicy<typename Candidate::Emulator_t, typename Candidate::Sensor_t, Logger> &&
     (Candidate::Id == Candidate::Emulator_t::Id) && (Candidate::Id == Candidate::Sensor_t::Id);
 
 } // namespace navkit::app_support